+++ src/main/java/com/starfireaviation/groundschool/events/EventStartedEvent.java
@@ -1,4 +1,4 @@
 package com.starfireaviation.groundschool.events;
 
-public record EventStartedEvent(int id) {
+public record EventStartedEvent(Long id) {
 }
