+++ src/main/java/com/starfireaviation/groundschool/quiz/QuizCompletedEvent.java
@@ -1,4 +1,4 @@
 package com.starfireaviation.groundschool.quiz;
 
-public record QuizCompletedEvent(int id) {
+public record QuizCompletedEvent(Long id) {
 }
